requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pillow>=9.0.0
opencv-python>=4.5.0